requirements.txt
pandas>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
torch>=2.0.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
numpy>=1.24.0
